+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 ### Added
 - #1561 Font editing - import ascent, descent, leading, kerning
+- Font editing - font name, ascent, descent, leading
 
 ### Fixed
 - #1623 Right side marker (gray line) in P-code